refactor listener into event sources, event sink and overall listener harness
